requirements.txt
alembic==1.4.1
astroid==2.3.3
autopep8==1.5.1
base58==2.0.0
Click==7.0
configparser==4.0.2
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
mccabe==0.6.1
PIC==0.1.0
Pillow==7.0.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
python-magic==0.4.15
six==1.13.0
SQLAlchemy==1.3.13
typed-ast==1.4.0
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
